Projeto escola - Form Mensalidade em manutenção

+++ Desenvolvimento-de-sistemas/Proj-Escola-BD/Proj-Escola-BD/frmMensalidade.cs
@@ -15,6 +15,7 @@ namespace Proj_Escola_BD
         DataTable DT;
         MySqlDataReader Sqlreader;
         Utilidades utilidades = new Utilidades();
+        double desconto, juros;
 
 
         public frmMensalidade()
@@ -24,21 +25,28 @@ namespace Proj_Escola_BD
 
         private void btnLimpar_Click(object sender, EventArgs e)
         {
-            txtMat.Text = "";
-            txtVPag.Text = "";
-            dtpPag.Value = DateTime.Now;
-            lbljurus.Text = "";
-            lblDesconto.Text = "";
+            campoMatricula.Text = "";
+            campoPagamento.Text = "";
+            campoData.Value = DateTime.Now;
+            outputJuros.Text = "";
+            outputDesconto.Text = "";
+            outputTotalPagamento.Text = "";
         }
 
         private void btnExcluir_Click(object sender, EventArgs e)
         {
+            // checa se o campo matricula está preenchido
+            if (!utilidades.validarCampo(campoMatricula.Text, "Matricula do aluno"))
+            {
+                return;
+            }
+
             cmdSql.Remove(0, cmdSql.Length);
             cmdSql.Append("delete from mensalidade");
             cmdSql.Append(" where Matricula_Alu = @Matricula_Alu");
 
             Conexao.Comandos.Parameters.Clear();
-            Conexao.Comandos.Parameters.AddWithValue("@Matricula_Alu", txtMat.Text);
+            Conexao.Comandos.Parameters.AddWithValue("@Matricula_Alu", campoMatricula.Text);
             Conexao.StrSql = cmdSql.ToString();
            
 
@@ -57,51 +65,43 @@ namespace Proj_Escola_BD
         private void btnIncluir_Click(object sender, EventArgs e)
         {
 
+            // valida as entradas necessarias
+            if (!utilidades.validarCampo(campoMatricula.Text, "Matricula da mensalidade"))
+            {
+                return;
+            }
+            if (!utilidades.validarCampo(campoPagamento.Text, "Valor a pagar"))
+            {
+                return;
+            }
+
+            // define o valor de juros ou desconto partindo do Dia do mês
+            calcularMudanca(campoData.Value.Day);
+
+            // cria o comando de inserir 
+            cmdSql.Remove(0, cmdSql.Length); // limpa a string de comando
+            cmdSql.Append("INSERT INTO mensalidade "); // insere na tabela mensalidade
+            cmdSql.Append("(Matricula_Alu ,DtPag_Mens , VlPag_Mens , Juros_Mens , Desconto_Mens) "); // insere nos campos
+            cmdSql.Append("Values "); // os seguintes valores
+            cmdSql.Append("(@Matricula_Alu, @DtPag_Mens , @VlPag_Mens , @Juros_Mens , @Desconto_Mens); "); // placeholders 
+
+            // tente executar o comando
             try
             {
-                if (!utilidades.validarCampo(txtMat.Text, "Matricula da mensalidade"))
-                {
-                    return;
-                }
-                if (!utilidades.validarCampo(txtVPag.Text, "Valor a pagar"))
-                {
-                    return;
-                }
+                Conexao.Comandos.Parameters.Clear(); // limpa os parametros anteriores
+                Conexao.Comandos.Parameters.AddWithValue("@Matricula_Alu", campoMatricula.Text); // adiciona a matricula
+                Conexao.Comandos.Parameters.AddWithValue("@DtPag_Mens", campoData.Value); // adiciona a data 
+                Conexao.Comandos.Parameters.AddWithValue("@VlPag_Mens", Double.Parse(campoPagamento.Text)); // adiciona o valor a ser pago
+                Conexao.Comandos.Parameters.AddWithValue("@Juros_Mens", this.juros); // adiciona o juros definido
+                Conexao.Comandos.Parameters.AddWithValue("@Desconto_Mens", this.desconto); // adiciona o desconto definido
+                Conexao.StrSql = cmdSql.ToString(); // envia a string de comando para o banco
 
-
-                calcularMudanca(dtpPag.Value.Day);
-
-
-                cmdSql.Remove(0, cmdSql.Length);
-                cmdSql.Append("insert into mensalidade ");
-                cmdSql.Append("(Matricula_Alu ,DtPag_Mens , VlPag_Mens , Juros_Mens , Desconto_Mens) ");
-                cmdSql.Append("Values ");
-                cmdSql.Append("(@Matricula_Alu, @DtPag_Mens , @VlPag_Mens , @Juros_Mens , @Desconto_Mens); ");
-
-                Conexao.Comandos.Parameters.Clear();
-                Conexao.Comandos.Parameters.AddWithValue("@Matricula_Alu", txtMat.Text);
-                Conexao.Comandos.Parameters.AddWithValue("@DtPag_Mens", dtpPag.Value);
-                Conexao.Comandos.Parameters.AddWithValue("@VlPag_Mens", Double.Parse(txtVPag.Text));
-                Conexao.Comandos.Parameters.AddWithValue("@Juros_Mens", Double.Parse(lbljurus.Text));
-                Conexao.Comandos.Parameters.AddWithValue("@Desconto_Mens", Double.Parse(lblDesconto.Text));
-
-                Conexao.StrSql = cmdSql.ToString();
-
-
-
+                // executa o comando
                 if (Conexao.executarComando() > 0)
                 {
-
-                    double valorAPagar = double.Parse(txtVPag.Text.ToString());
-                    double juros = double.Parse(lbljurus.Text.ToString());
-                    double desc = double.Parse(lblDesconto.Text.ToString());
-
-                    exibirTotal(calcularTotal(valorAPagar, juros, desc));
+                    exibirTotal(calcularTotal());
                     ChamarGrid();
-
                     MessageBox.Show("Inclusão com sucesso");
-
-
                 }
                 else
                 {
@@ -110,7 +110,6 @@ namespace Proj_Escola_BD
             }
             catch (Exception ex)
             {
-                MessageBox.Show("A matricula precisa ser igual a do aluno");
                 MessageBox.Show(ex.Message);
             }
             
@@ -118,202 +117,175 @@ namespace Proj_Escola_BD
 
         private void btnConsultar_Click(object sender, EventArgs e)
         {
-
-          
-
-            calcularMudanca(dtpPag.Value.Day);
-            
+            if (!utilidades.validarCampo(campoMatricula.Text, "Matricula do aluno"))
+            {
+                return;
+            }
 
             cmdSql.Remove(0, cmdSql.Length);
-            cmdSql.Append("select * ");
-            cmdSql.Append("from mensalidade");
-            cmdSql.Append(" WHERE Matricula_Alu = @Matricula_Alu");
+            cmdSql.Append("SELECT ");
+            cmdSql.Append("Matricula_Alu as 'Matricula', ");
+            cmdSql.Append("DtPag_Mens as 'Data de Pagamento', ");
+            cmdSql.Append("VlPag_Mens as 'Valor da Mensalidade', ");
+            cmdSql.Append("Juros_Mens as 'Juros Aplicado', ");
+            cmdSql.Append("Desconto_Mens as 'Desconto Aplicado', ");
+            cmdSql.Append("VlPag_Mens+(VlPag_Mens*(Juros_Mens/100)) as 'Valor com Juros', ");
+            cmdSql.Append("VlPag_Mens-(VlPag_Mens*(Desconto_Mens/100)) as 'Valor com Desconto' ");
+            cmdSql.Append("FROM mensalidade WHERE Matricula_Alu = @matricula");
 
             Conexao.Comandos.Parameters.Clear();
-            Conexao.Comandos.Parameters.AddWithValue("@Matricula_Alu", txtMat.Text);
-
+            Conexao.Comandos.Parameters.AddWithValue("@matricula", campoMatricula.Text.ToString());
             Conexao.StrSql = cmdSql.ToString();
-
-
-
-            Sqlreader = Conexao.retornarDataReader();
-
-            if (Sqlreader.Read())
-            {
-              
-                txtVPag.Text = Sqlreader["VlPag_Mens"].ToString();
-                dtpPag.Text = Sqlreader["DtPag_Mens"].ToString();
-                lbljurus.Text = Sqlreader["Juros_Mens"].ToString();
-                lblDesconto.Text = Sqlreader["Desconto_Mens"].ToString();
-
-                double valorAPagar = double.Parse(txtVPag.Text.ToString());
-                double juros = double.Parse(lbljurus.Text.ToString());
-                double desc = double.Parse(lblDesconto.Text.ToString());
-                exibirTotal(calcularTotal(valorAPagar, juros, desc));
-                ChamarGrid();
-
-
-            }
-            else
-            {
-                MessageBox.Show("Não localizado");
-            }
+            conjuntoDeDados = Conexao.RetornarDataSet();
+            DT = conjuntoDeDados.Tables[0];
+            gridMensalidade.DataSource = DT;
         }
         
 
         private void btnPesquisar_Click(object sender, EventArgs e)
         {
-            cmdSql.Remove(0, cmdSql.Length);
-            cmdSql.Append("select *, VlPag_Mens+(VlPag_Mens*(Juros_Mens/100)) as Total_Juros, VlPag_Mens-(VlPag_Mens*(Desconto_Mens/100)) as Total_Desc ");
-            cmdSql.Append("from mensalidade");
-
-            Conexao.StrSql = cmdSql.ToString();
-            conjuntoDeDados = Conexao.RetornarDataSet();
-
-            DT = conjuntoDeDados.Tables[0];
-            gridMens.DataSource = DT;
+            ChamarGrid();
         }
 
         private void btnAlterar_Click(object sender, EventArgs e)
         {
-            if (!utilidades.validarCampo(txtMat.Text, "Matricula do aluno"))
+
+            if (!utilidades.validarCampo(campoMatricula.Text, "Matricula do aluno"))
             {
                 return;
             }
-            if (!utilidades.validarCampo(txtVPag.Text, "Valor a pagar"))
+            if (!utilidades.validarCampo(campoPagamento.Text, "Valor a pagar"))
             {
                 return;
             }
 
-            double valorAPagar = double.Parse(txtVPag.Text.ToString());
-            double juros = double.Parse(lbljurus.Text.ToString());
-            double desc = double.Parse(lblDesconto.Text.ToString());
+            calcularMudanca(campoData.Value.Day);
+            exibirTotal(calcularTotal());
 
-            calcularMudanca(dtpPag.Value.Day);
-            exibirTotal(calcularTotal(valorAPagar, juros, desc));
-
-            cmdSql.Remove(0, cmdSql.Length);
-            cmdSql.Append("update mensalidade set ");
-            cmdSql.Append("DtPag_Mens =  @DtPag_Mens , VlPag_Mens = @VlPag_Mens , Juros_Mens = @Juros_Mens  , Desconto_Mens = @Desconto_Mens");
-            cmdSql.Append(" where Matricula_Alu= @Matricula_Alu");
-            //update filme set nome_filme='Teste',data_lanc_filme = curdate() where id_filme = 5;
-
-
-            Conexao.Comandos.Parameters.Clear();
-            Conexao.Comandos.Parameters.AddWithValue("@Matricula_Alu", txtMat.Text);
-            Conexao.Comandos.Parameters.AddWithValue("@DtPag_Mens", dtpPag.Value);
-            Conexao.Comandos.Parameters.AddWithValue("@VlPag_Mens", Double.Parse(txtVPag.Text));
-            Conexao.Comandos.Parameters.AddWithValue("@Juros_Mens", Double.Parse(lbljurus.Text));
-            Conexao.Comandos.Parameters.AddWithValue("@Desconto_Mens", Double.Parse(lblDesconto.Text));
-
-            
-            Conexao.StrSql = cmdSql.ToString();
-            
-
-
-
-            if (Conexao.executarComando() > 0)
+            try
             {
-                ChamarGrid();
-                MessageBox.Show("Alteração com sucesso");
-                
+                cmdSql.Remove(0, cmdSql.Length);
+                cmdSql.Append("update mensalidade set ");
+                cmdSql.Append("DtPag_Mens =  @DtPag_Mens , VlPag_Mens = @VlPag_Mens , Juros_Mens = @Juros_Mens  , Desconto_Mens = @Desconto_Mens");
+                cmdSql.Append(" where Matricula_Alu= @Matricula_Alu");
+
+                Conexao.Comandos.Parameters.Clear();
+                Conexao.Comandos.Parameters.AddWithValue("@Matricula_Alu", campoMatricula.Text);
+                Conexao.Comandos.Parameters.AddWithValue("@DtPag_Mens", campoData.Value);
+                Conexao.Comandos.Parameters.AddWithValue("@VlPag_Mens", Double.Parse(campoPagamento.Text));
+                Conexao.Comandos.Parameters.AddWithValue("@Juros_Mens", this.juros);
+                Conexao.Comandos.Parameters.AddWithValue("@Desconto_Mens", this.desconto);
+
+                Conexao.StrSql = cmdSql.ToString();
+                if (Conexao.executarComando() > 0)
+                {
+                    ChamarGrid();
+                    MessageBox.Show("Alteração com sucesso");
+
+                }
+                else
+                {
+                    MessageBox.Show("Erro na Alteração");
+                }
             }
-            else
+            catch (Exception ex)
             {
-                MessageBox.Show("Erro na Alteração");
+                MessageBox.Show(ex.Message);
             }
-        }
-
-        private void lbljurus_Click(object sender, EventArgs e)
-        {
-
         }
 
         private void calcularMudanca (int dataPagamento)
         {
+            // calcular o dia do pagamento e define as variaveis de desconto e de juros se houver
+            this.juros = 0;
+            this.desconto = 0;
 
             if (dataPagamento > 5)
             {
-
-                double jurus = 7.25;
-                lbljurus.Text = "7,25";
-
-                double desc = 0;
-                lblDesconto.Text = "0";
-
+                this.juros = 7.25;
             }
             else if (dataPagamento < 5)
             {
-                double desc = 5;
-                double jurus = 0;
-                lblDesconto.Text = "5";
-                lbljurus.Text = "0";
-
-
+                this.desconto = 5;
             }
-            else
-            {
-                double desc = 0;
-                double jurus = 0;
-                lblDesconto.Text = "0";
-                lbljurus.Text = "0";
-            }
+            
         }
 
-        private double calcularTotal (double valorApagar, double jurosCal, double descCal )
+        private double calcularTotal ()
         {
-            valorApagar = Double.Parse(txtVPag.Text);
-            jurosCal = Double.Parse(lbljurus.Text);
-            descCal = Double.Parse(lblDesconto.Text);
-
-            double total;
-
-            if (descCal > 0 && jurosCal == 0)
+            double valorApagar = Double.Parse(campoPagamento.Text);
+            // não teve desconto nem juros
+            if (this.desconto == 0 && this.juros == 0)
             {
-                total =  valorApagar - (valorApagar * (descCal / 100));
-                return total;
+                return valorApagar;
             }
-            else if (jurosCal > 0 && descCal == 0)
+            // teve desconto
+            else if (this.desconto > 0)
             {
-                total = valorApagar + (valorApagar * (jurosCal / 100));
-                return total;
+                return valorApagar - (valorApagar * (this.desconto / 100));
             }
+            // teve juros
+            else if (this.juros > 0)
+            {
+                return valorApagar + (valorApagar * (this.juros / 100));
+            }
+
+            // apenas para evitar erro de tipagem, esse else nunca vai ser executado
             else
             {
-                total = valorApagar;
-                return total;
+                return valorApagar;
             }
+
+
         }
 
         private void exibirTotal (double total)
         {
-            lblTotal.Text = total.ToString();
+            outputDesconto.Text = this.desconto + "%";
+            outputJuros.Text = this.juros + "%";
+            outputTotalPagamento.Text = "R$: " + total;
         }
 
-        private double converterValores ()
+        private void selecionarMensalidade(object sender, DataGridViewCellEventArgs e)
         {
-            double valorAPagar = double.Parse(txtVPag.Text.ToString());
-            double juros = double.Parse(lbljurus.Text.ToString());
-            double desc = double.Parse(lblDesconto.Text.ToString());
-            return valorAPagar;
-            return juros;
-            return desc;
+            
+            campoMatricula.Text = gridMensalidade.CurrentRow.Cells[0].Value.ToString();
+            campoData.Text = gridMensalidade.CurrentRow.Cells[1].Value.ToString();
+            campoPagamento.Text = gridMensalidade.CurrentRow.Cells[2].Value.ToString();
+            outputJuros.Text = gridMensalidade.CurrentRow.Cells[3].Value.ToString();
+            outputDesconto.Text = gridMensalidade.CurrentRow.Cells[4].Value.ToString();
+            // checa se o valor real a ser pago é o do juros ou do desconto
+            if (Double.Parse(outputJuros.Text) == 0)
+            {
+                outputTotalPagamento.Text = "R$: " + gridMensalidade.CurrentRow.Cells[6].Value.ToString();
+            }
+
+            if (Double.Parse(outputDesconto.Text) == 0)
+            {
+                outputTotalPagamento.Text = "R$: " + gridMensalidade.CurrentRow.Cells[5].Value.ToString();
+            }
+
+            outputDesconto.Text += "%";
+            outputJuros.Text += "%";
         }
 
         private void ChamarGrid()
         {
- 
-            
-                cmdSql.Remove(0, cmdSql.Length);
-                cmdSql.Append("select *, VlPag_Mens+(VlPag_Mens*(Juros_Mens/100)) as Total_Juros, VlPag_Mens-(VlPag_Mens*(Desconto_Mens/100)) as Total_Desc ");
-                cmdSql.Append("from mensalidade");
+            cmdSql.Remove(0, cmdSql.Length);
+            cmdSql.Append("SELECT ");
+            cmdSql.Append("Matricula_Alu as 'Matricula', ");
+            cmdSql.Append("DtPag_Mens as 'Data de Pagamento', ");
+            cmdSql.Append("VlPag_Mens as 'Valor da Mensalidade', ");
+            cmdSql.Append("Juros_Mens as 'Juros Aplicado', ");
+            cmdSql.Append("Desconto_Mens as 'Desconto Aplicado', ");
+            cmdSql.Append("VlPag_Mens+(VlPag_Mens*(Juros_Mens/100)) as 'Valor com Juros', ");
+            cmdSql.Append("VlPag_Mens-(VlPag_Mens*(Desconto_Mens/100)) as 'Valor com Desconto' ");
+            cmdSql.Append("FROM mensalidade");
+            Conexao.StrSql = cmdSql.ToString();
+            conjuntoDeDados = Conexao.RetornarDataSet();
+            DT = conjuntoDeDados.Tables[0];
+            gridMensalidade.DataSource = DT;
 
-                Conexao.StrSql = cmdSql.ToString();
-                conjuntoDeDados = Conexao.RetornarDataSet();
-
-                DT = conjuntoDeDados.Tables[0];
-                gridMens.DataSource = DT;
-           
         }
     }
 }
